requirements.txt
# Core dependencies
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.10
googletrans>=4.0.0
requests>=2.31.0
flask>=2.3.0
click>=8.0.0
reportlab>=4.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0